requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
babel==2.17.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
diff-match-patch==20241021
distlib==0.4.0
Django==5.2.5
django-autoslug==1.9.9
django-celery-beat==2.8.1
django-constance==4.3.2
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-environ==0.12.0
django-filter==25.1
django-guardian==3.1.0
django-import-export==4.3.9
django-modeltranslation==0.19.16
django-money==3.5.4
django-phonenumber-field==8.1.0
django-simple-history==3.10.1
django-timezone-field==7.1
django-unfold==0.65.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
filelock==3.19.1
googleapis-common-protos==1.70.0
grpcio==1.74.0
identify==2.6.14
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
Markdown==3.9
nodeenv==1.9.1
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-celery==0.57b0
opentelemetry-instrumentation-django==0.57b0
opentelemetry-instrumentation-logging==0.57b0
opentelemetry-instrumentation-requests==0.57b0
opentelemetry-instrumentation-wsgi==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
pre_commit==4.3.0
prompt_toolkit==3.0.52
protobuf==6.32.0
psycopg==3.2.9
psycopg-binary==3.2.10
py-moneyed==3.0
PyJWT==2.10.1
pyotp==2.9.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pyupgrade==3.20.0
PyYAML==6.0.2
qrcode==8.2
redis==6.4.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tokenize_rt==6.2.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.34.0
wcwidth==0.2.13
wrapt==1.17.3
zipp==3.23.0
